+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.2.0] - 2026-04-14
+
+### Added
+- Support for negation via `!` for all terms in CLI
+  - Ex. `#followup !#archive`, `meeting !title:draft`
+
 ## [0.1.0] - 2026-04-08
 
 First tagged prerelease. See [README.md](README.md) and [the cli reference](docs/cli-reference.md) for more information.